Refactor DdsDataset extension methods to proper location
Move DdsDataset query and metadata methods from query.rs to the main DdsDataset implementation block in dds.rs, along with their corresponding tests. This follows proper Rust organization patterns by keeping implementation methods with their types rather than scattered across modules.

Changes:
- Move query(), list_variables(), list_coordinates(), get_variable_info(), get_coordinate_info(), has_variable(), has_coordinate() methods to dds.rs
- Move corresponding tests to dds.rs test module
- Remove unused HashSet import from query.rs
- Maintain full test coverage and functionality
